more warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@14728 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Sessions/Src/Utils.cpp
@@ -50,7 +50,7 @@ void AddSessionMark(MCONTACT hContact, int mode, char bit)
 			else pszBuffer = szValue.detach();
 
 			char temp_1 = pszBuffer[0];
-			for (size_t i = 0; i < g_ses_count; i++) {
+			for (int i = 0; i < g_ses_count; i++) {
 				char temp_2 = pszBuffer[i + 1];
 				pszBuffer[i + 1] = temp_1;
 				temp_1 = temp_2;
